feat: refactor viewer as separate app
