+++ viewer/__init__.py
@@ -1,0 +1,4 @@
+"""Main TUI for az-firewall-watch."""
+from .app import FirewallLogApp
+
+__all__ = ["FirewallLogApp"]
